Correciones a facturacion

+++ src/com/juanyjulian/facturacion/dominio/Factura.java
@@ -3,12 +3,12 @@ package com.juanyjulian.facturacion.dominio;
 import java.time.LocalDate;
 
 public abstract class Factura {
-    protected long valor;
+    protected double valor;
     protected String cliente;
     protected LocalDate fecha;
 
-    public Factura(long valor, String cliente, LocalDate fecha) {
+    public Factura(double valor, String cliente, LocalDate fecha) {
     }
 
-    protected abstract long calcularValor();
+    protected abstract double calcularValor();
 }
